separate utils and constants into different files
