+++ .gitignore
@@ -109,6 +109,7 @@ venv/
 ENV/
 env.bak/
 venv.bak/
+env.yml
 
 # Spyder project settings
 .spyderproject
